refactor(java): fold static serializer naming into registry
